style: move bytes extensions from Unsafe to ByteExtensions
