+++ tests/Neo.Extensions.Tests/UT_ByteExtensions.cs
@@ -32,7 +32,7 @@ namespace Neo.Extensions.Tests
             Assert.AreEqual("", empty.ToHexString(false));
             Assert.AreEqual("", empty.ToHexString(true));
 
-            byte[] str1 = new byte[] { (byte)'n', (byte)'e', (byte)'o' };
+            byte[] str1 = [(byte)'n', (byte)'e', (byte)'o'];
             Assert.AreEqual("6e656f", str1.ToHexString());
             Assert.AreEqual("6e656f", str1.ToHexString(false));
             Assert.AreEqual("6f656e", str1.ToHexString(true));
@@ -58,15 +58,46 @@ namespace Neo.Extensions.Tests
             result = span.ToHexString();
             Assert.AreEqual(0, result.Length);
 
-            input = new byte[] { 0x5A };
+            input = [0x5A];
             span = new ReadOnlySpan<byte>(input);
             result = span.ToHexString();
             Assert.AreEqual("5a", result);
 
-            input = new byte[] { 0x01, 0x23, 0x45, 0x67, 0x89, 0xAB, 0xCD, 0xEF };
+            input = [0x01, 0x23, 0x45, 0x67, 0x89, 0xAB, 0xCD, 0xEF];
             span = new ReadOnlySpan<byte>(input);
             result = span.ToHexString();
             Assert.AreEqual("0123456789abcdef", result);
+        }
+
+        [TestMethod]
+        public void TestNotZero()
+        {
+            Assert.IsFalse(new ReadOnlySpan<byte>(Array.Empty<byte>()).NotZero());
+            Assert.IsFalse(new ReadOnlySpan<byte>(new byte[4]).NotZero());
+            Assert.IsFalse(new ReadOnlySpan<byte>(new byte[7]).NotZero());
+            Assert.IsFalse(new ReadOnlySpan<byte>(new byte[8]).NotZero());
+            Assert.IsFalse(new ReadOnlySpan<byte>(new byte[9]).NotZero());
+            Assert.IsFalse(new ReadOnlySpan<byte>(new byte[11]).NotZero());
+
+            Assert.IsTrue(new ReadOnlySpan<byte>([0x00, 0x00, 0x00, 0x01]).NotZero());
+            Assert.IsTrue(new ReadOnlySpan<byte>([0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01]).NotZero());
+            Assert.IsTrue(new ReadOnlySpan<byte>([0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01]).NotZero());
+            Assert.IsTrue(new ReadOnlySpan<byte>([0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00]).NotZero());
+            Assert.IsTrue(new ReadOnlySpan<byte>([0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01]).NotZero());
+
+            var bytes = new byte[64];
+            for (int i = 0; i < bytes.Length; i++)
+            {
+                ReadOnlySpan<byte> span = bytes.AsSpan();
+                Assert.IsFalse(span[i..].NotZero());
+
+                for (int j = i; j < bytes.Length; j++)
+                {
+                    bytes[j] = 0x01;
+                    Assert.IsTrue(span[i..].NotZero());
+                    bytes[j] = 0x00;
+                }
+            }
         }
     }
 }
